Extract email templates into separate files (#335)
* Extracted the welcome mail into a separate folder

* Wrong use of context in django and incorrect dictionary syntax corrected

* Moved template into new folder

* Renamed a variable to be more explanatory and conform to previous naming

* Moved mails into separate files, and uses template rendering instead of formatted strings

* Wrong template path, corrected it

* Moved send_payment_mail into mail.py. Alos, template used wrong value before.

* Reduced send_payment_mail function to use send_mail.

* Corrected formatting in file unrelated to this PR to satisfy black
