+++ stregsystem/templates/mail/deposit_manual.html
@@ -1,0 +1,26 @@
+<html>
+    <head></head>
+    <body>
+        <p>
+            Hej {{ firstname }}!<br><br>
+            Vi har med stort besvær indsat pokkers {{ formatted_amount }} stregdollars på din konto: "{{ username }}". <br><br>
+            Da du ikke skrev dit brugernavn korrekt, men i stedet skrev '{{ mobilepay_comment }}' var de stakkels TREOer desværre nødt til at tage flere minutter ud af deres dag for at indsætte dine penge manuelt.
+            Vil du nyde godt af vores automatiske indbetaling kan du i fremtiden med fordel skrive dit brugernavn korrekt i MobilePay kommentaren: '{{ username }}'.
+            Udnytter du vores QR-kode generator klarer den også denne komplicerede process for dig.
+            
+            Hvis du ikke ønsker at modtage flere mails som denne kan du skrive en mail til: <a href="mailto:treo@fklub.dk?Subject=Klage" target="_top">treo@fklub.dk</a><br><br>
+            Mvh,<br>
+            TREOen<br>
+            ====================================================================<br>
+            Hello {{ firstname }}!<br><br>
+            We've had great trouble inserting {{ formatted_amount }} stregdollars on your account: "{{ username }}". <br><br>
+            This is due to you not you not writing your username correctly, and instead writing '{{ mobilepay_comment }}'. The poor TREOs had to take multiple minutes out of their day to insert your money manually.
+            If you want to utilise our automatic fill-up in future, you can write your username correctly in the MobilePay comment: '{{ username }}'
+            Our QR-code generator also handles this very complicated process for you. 
+            
+            If you do not desire to receive any more mails of this sort, please file a complaint to: <a href="mailto:treo@fklub.dk?Subject=Klage" target="_top">treo@fklub.dk</a><br><br>
+            Kind regards,<br>
+            TREOen
+        </p>
+    </body>
+</html>
